third_party/WebKit/LayoutTests/fast/canvas/webgl/texImage-imageBitmap-from-image-resize.html

<!DOCTYPE html>
<html>
<head>
<script src="./resources/webgl-test-utils-full.js"></script>
<script src="./resources/tex-image-and-sub-image-utils.js"></script>
<script src="./resources/tex-image-and-sub-image-image-bitmap-utils.js"></script>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
</head>
<body>
<script>
var wtu = WebGLTestUtils;
var tiu = TexImageUtils;
var gl = null;
var internalFormat = "RGBA";
var pixelFormat = "RGBA";
var pixelType = "UNSIGNED_BYTE";
var redColor = [255, 0, 0];
var greenColor = [0, 255, 0];
var blackColor = [0, 0, 0];
var darkRed = [26, 0, 0];
var darkGreen = [0, 26, 0];

var image1 = new Image();
var image2 = new Image();

function generateTest()
{
    var bitmaps = [];

    var canvas = document.createElement('canvas');
    canvas.width = 32;
    canvas.height = 32;
    document.body.appendChild(canvas);
    gl = canvas.getContext("webgl");

    gl.clearColor(0,0,0,1);
    gl.clearDepth(1);

    var p1 = createImageBitmap(image1, {resizeWidth: 4, resizeHeight: 4, resizeQuality: "high"}).then(function(imageBitmap) { bitmaps.defaultOption = imageBitmap });
    var p2 = createImageBitmap(image1, {imageOrientation: "none", premultiplyAlpha: "premultiply", resizeWidth: 4, resizeHeight: 4, resizeQuality: "high"}).then(function(imageBitmap) { bitmaps.noFlipYPremul = imageBitmap });
    var p3 = createImageBitmap(image1, {imageOrientation: "none", premultiplyAlpha: "default", resizeWidth: 4, resizeHeight: 4, resizeQuality: "high"}).then(function(imageBitmap) { bitmaps.noFlipYDefault = imageBitmap });
    var p4 = createImageBitmap(image1, {imageOrientation: "none", premultiplyAlpha: "none", resizeWidth: 4, resizeHeight: 4, resizeQuality: "high"}).then(function(imageBitmap) { bitmaps.noFlipYUnpremul = imageBitmap });
    var p5 = createImageBitmap(image1, {imageOrientation: "flipY", premultiplyAlpha: "premultiply", resizeWidth: 4, resizeHeight: 4, resizeQuality: "high"}).then(function(imageBitmap) { bitmaps.flipYPremul = imageBitmap });
    var p6 = createImageBitmap(image1, {imageOrientation: "flipY", premultiplyAlpha: "default", resizeWidth: 4, resizeHeight: 4, resizeQuality: "high"}).then(function(imageBitmap) { bitmaps.flipYDefault = imageBitmap });
    var p7 = createImageBitmap(image1, {imageOrientation: "flipY", premultiplyAlpha: "none", resizeWidth: 4, resizeHeight: 4, resizeQuality: "high"}).then(function(imageBitmap) { bitmaps.flipYUnpremul = imageBitmap });
    var p8 = createImageBitmap(image2, {resizeWidth: 4, resizeHeight: 4, resizeQuality: "high"}).then(function(imageBitmap) { bitmaps.colorSpaceDef = imageBitmap });
    var p9 = createImageBitmap(image2, {colorSpaceConversion: "none", resizeWidth: 4, resizeHeight: 4, resizeQuality: "high"}).then(function(imageBitmap) { bitmaps.colorSpaceNone = imageBitmap });
    var p10 = createImageBitmap(image2, {colorSpaceConversion: "default", resizeWidth: 4, resizeHeight: 4, resizeQuality: "high"}).then(function(imageBitmap) { bitmaps.colorSpaceDefault = imageBitmap });
    return Promise.all([p1, p2, p3, p4, p5, p6, p7, p8, p9, p10]).then(function() {
        var alphaVal = 0.5;
        var testPassed = runTest(bitmaps, alphaVal, true);
        if (!testPassed)
            assert_true(false, 'Test failed');
    }, function() {
        assert_true(false, 'Promise rejected');
    });
}

function loadImage(image, src) {
    return new Promise((resolve, reject) => {
        image.src = src;
        image.addEventListener('load', () => {
            resolve();
        });
    });
}

promise_test(function() {
    Promise.all([loadImage(image1, 'resources/red-green-semi-transparent.png'), loadImage(image2, 'resources/square-with-colorspin-profile.png')]).then(generateTest);
}, 'createImageBitmap(HTMLImageElement) with resize and other options');
</script>
</body>
</html>
